change ui error handler
