+++ lib/lenra_widget.dart
@@ -1,52 +1,118 @@
 import 'package:flutter/material.dart';
 import 'package:lenra_components/lenra_components.dart';
+import 'package:lenra_ui_runner/lenra_component_builder.dart';
 import 'package:lenra_ui_runner/widget_model.dart';
 import 'package:provider/provider.dart';
+import 'package:lenra_ui_runner/components/lenra_button.dart';
+import 'package:lenra_ui_runner/components/lenra_checkbox.dart';
+import 'package:lenra_ui_runner/components/lenra_icon.dart';
+import 'package:lenra_ui_runner/components/lenra_image.dart';
+import 'package:lenra_ui_runner/components/lenra_overlay_entry.dart';
+import 'package:lenra_ui_runner/components/lenra_radio.dart';
+import 'package:lenra_ui_runner/components/lenra_text.dart';
+import 'package:lenra_ui_runner/components/lenra_textfield.dart';
+import 'package:lenra_ui_runner/props_parser.dart';
+import 'package:lenra_ui_runner/components/lenra_dropdown_button.dart';
+import 'package:lenra_ui_runner/components/lenra_flexible.dart';
+import 'package:lenra_ui_runner/components/lenra_container.dart';
+import 'package:lenra_ui_runner/components/lenra_stack.dart';
+import 'package:lenra_ui_runner/components/lenra_slider.dart';
+import 'package:lenra_ui_runner/components/lenra_actionable.dart';
+import 'package:lenra_ui_runner/components/lenra_menu.dart';
+import 'package:lenra_ui_runner/components/lenra_menu_item.dart';
+import 'package:lenra_ui_runner/components/lenra_toggle.dart';
+import 'package:lenra_ui_runner/components/lenra_status_sticker.dart';
+import 'package:lenra_ui_runner/components/lenra_flex.dart';
+import 'package:lenra_ui_runner/components/lenra_wrap.dart';
 
 class LenraWidget extends StatelessWidget {
+  static Function appErrorUI = () {};
+  static List<Widget> _errors = [];
+  static Function appErrorCallback = () {};
+
   @override
   Widget build(BuildContext context) {
-    WidgetModel widgetModel = context.watch<WidgetModel>();
-    Map<String, dynamic> ui = widgetModel.ui;
-    bool error = widgetModel.error;
+    Map<String, dynamic> ui = context.select<WidgetModel, Map<String, dynamic>>((model) => model.ui);
+
+    bool error = context.select<WidgetModel, bool>((model) => model.error);
 
     if (error) {
-      WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            duration: Duration(minutes: 2),
-            backgroundColor: LenraColorThemeData.lenraCustomRed,
-            content: LenraFlex(
-              spacing: 2,
-              direction: Axis.horizontal,
-              children: widgetModel.errors +
-                  [
-                    Spacer(),
-                    //TODO: We should consider using api route to notify dev
-                    LenraButton(
-                      disabled: true,
-                      type: LenraComponentType.secondary,
-                      onPressed: () {},
-                      text: "Contact developper",
-                    ),
-                    LenraButton(
-                      type: LenraComponentType.secondary,
-                      onPressed: () {
-                        ScaffoldMessenger.of(context).clearSnackBars();
-                      },
-                      leftIcon: Icon(Icons.close),
-                    )
-                  ],
-            ),
-          ),
-        );
-      });
+      List<Widget> error = context.select<WidgetModel, List<Widget>>((model) => model.errors);
+      LenraWidget._errors = error;
+      appErrorCallback(context, error);
     }
 
     if (ui.keys.contains("root") && ui.keys.length == 1) {
-      return WidgetModel.parseJson(ui["root"]);
+      return LenraWidget.parseJson(ui["root"]);
     }
 
-    return WidgetModel.parseJson(ui);
+    return LenraWidget.parseJson(ui);
+  }
+
+  static final Map<String, LenraComponentBuilder> componentsMapping = {
+    'text': LenraTextBuilder(),
+    'textfield': LenraTextfieldBuilder(),
+    'button': LenraButtonBuilder(),
+    'checkbox': LenraCheckboxBuilder(),
+    'image': LenraImageBuilder(),
+    'radio': LenraRadioBuilder(),
+    'menu': LenraMenuBuilder(),
+    'menuItem': LenraMenuItemBuilder(),
+    'toggle': LenraToggleBuilder(),
+    'statusSticker': LenraStatusStickerBuilder(),
+    'flex': LenraFlexBuilder(),
+    'container': LenraContainerBuilder(),
+    'actionable': LenraActionableBuilder(),
+    'dropdownButton': LenraDropdownButtonBuilder(),
+    'flexible': LenraFlexibleBuilder(),
+    'wrap': LenraWrapBuilder(),
+    'stack': LenraStackBuilder(),
+    'slider': LenraSliderBuilder(),
+    'overlayEntry': LenraOverlayEntryBuilder(),
+    'icon': LenraIconBuilder(),
+  };
+
+  static Widget parseJson(Map<String, dynamic> json) {
+    if (json.isEmpty) return appErrorUI(_errors);
+
+    String? type = getType(json);
+    if (type == null) {
+      throw "No type in component. It should never happen";
+    }
+    LenraComponentBuilder? builder = getComponentBuilder(type);
+    if (builder == null) throw "Componnent mapping does not handle type $type";
+    Map<Symbol, dynamic> parsedProps = parseProps(json, builder.propsTypes);
+    return builder.build(parsedProps);
+  }
+
+  static String? getType(Map<String, dynamic> json) {
+    return json['type'] as String?;
+  }
+
+  static LenraComponentBuilder? getComponentBuilder(String type) {
+    if (componentsMapping.containsKey(type)) {
+      return componentsMapping[type];
+    }
+
+    return null;
+  }
+
+  static Map<Symbol, dynamic> parseProps(Map<String, dynamic> json, Map<String, Type> propsTypes) {
+    Map<Symbol, dynamic> transformedProps = {};
+
+    json.forEach((key, value) {
+      if (propsTypes.containsKey(key)) {
+        Function? parser = getParser(propsTypes[key]!);
+        if (parser != null) {
+          transformedProps[Symbol(key)] = parser(value);
+        }
+      }
+    });
+
+    return transformedProps;
+  }
+
+  static Function? getParser(Type t) {
+    return ParserExt.typeParsers[t];
   }
 }
