+++ lib/props_parser.dart
@@ -7,6 +7,7 @@ import 'package:lenra_components/theme/lenra_slider_style.dart';
 import 'package:lenra_components/theme/lenra_theme_data.dart';
 import 'package:lenra_components/theme/lenra_toggle_style.dart';
 import 'package:lenra_ui_runner/components/listeners/listener.dart' as lenra;
+import 'package:lenra_ui_runner/lenra_widget.dart';
 import 'package:lenra_ui_runner/widget_model.dart';
 import 'package:lenra_ui_runner/utils/icon_util.dart';
 import 'package:lenra_ui_runner/components/children_widgets.dart';
@@ -596,7 +597,7 @@ class Parser {
   }
 
   static Widget parseWidget(Map<String, dynamic> props) {
-    return WidgetModel.parseJson(props);
+    return LenraWidget.parseJson(props);
   }
 
   static FilterQuality parseFilterQuality(String value) {
